fix: refactor of TableCells component
